+++ internal/domain/models/users.go
@@ -12,13 +12,3 @@ type UserLogin struct {
 	Email string `json:"email" validate:"email"`
 	Pass  string `json:"pass" validate:"min=8"`
 }
-
-type Book struct {
-	BookID      string `json:"book_id,omitempty"`
-	Author      string `json:"author"`
-	Lable       string `json:"lable"`
-	Description string `json:"description"`
-	Genre       string `json:"genre"`
-	WritedAt    string `json:"writed_at"`
-	Count       int    `json:"count,omitempty"`
-}
