+++ internal/repository/db-storage/storage.go
@@ -2,10 +2,11 @@ package dbstorage
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"log"
-	"time"
 
-	"github.com/Dorrrke/library0706/internal/domain/models"
+	"github.com/golang-migrate/migrate/v4"
 	"github.com/jackc/pgx/v5"
 )
 
@@ -23,99 +24,20 @@ func NewStorage(connStr string) (*Storage, error) {
 	}, nil
 }
 
-func (s *Storage) GetBooksList() ([]models.Book, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	rows, err := s.conn.Query(ctx, "SELECT * FROM books")
+func Migrations(dbDsn string, migratePath string) error {
+	mPath := fmt.Sprintf("file://%s", migratePath)
+	m, err := migrate.New(mPath, dbDsn)
 	if err != nil {
-		return nil, err
+		return err
 	}
 
-	var books []models.Book
-	for rows.Next() {
-		var book models.Book
-		err := rows.Scan(&book.BookID, &book.Author, &book.Lable, &book.Description, &book.Genre, &book.WritedAt, &book.Count)
-		if err != nil {
-			return nil, err
+	if err := m.Up(); err != nil {
+		if errors.Is(err, migrate.ErrNoChange) {
+			log.Println("Database is up to date")
 		}
-		books = append(books, book)
-	}
-	return books, nil
-}
-
-func (s *Storage) SaveBook(book models.Book) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.conn.Exec(ctx, "INSERT INTO books (bid, author, lable, description, genre, writed_at, count) VALUES ($1, $2, $3, $4, $5, $6, $7)",
-		book.BookID, book.Author, book.Lable, book.Description, book.Genre, book.WritedAt, book.Count,
-	)
-
-	if err != nil {
 		return err
 	}
 
-	return nil
-}
-
-func (s *Storage) SaveBooks(books []models.Book) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	tx, err := s.conn.Begin(ctx)
-	if err != nil {
-		return err
-	}
-
-	defer func() {
-		err := tx.Rollback(ctx)
-		if err != nil {
-			log.Println("Error rollback: ", err.Error())
-		}
-	}()
-
-	_, err = tx.Prepare(ctx, "save_book", "INSERT INTO books (bid, author, lable, description, genre, writed_at, count) VALUES ($1, $2, $3, $4, $5, $6, $7)")
-	if err != nil {
-		return err
-	}
-
-	for _, book := range books {
-		_, err = tx.Exec(ctx, "save_book", book.BookID, book.Author, book.Lable, book.Description, book.Genre, book.WritedAt, book.Count)
-		if err != nil {
-			return err
-		}
-	}
-
-	return tx.Commit(ctx)
-}
-
-func (s *Storage) GetUser(email string) (models.User, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	var user models.User
-
-	row := s.conn.QueryRow(ctx, "SELECT * FROM users WHERE email = $1", email)
-
-	err := row.Scan(&user.UID, &user.Name, &user.Age, &user.Email, &user.Pass)
-	if err != nil {
-		return models.User{}, err
-	}
-
-	return user, nil
-}
-
-func (s *Storage) SaveUser(user models.User) error {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	defer cancel()
-
-	_, err := s.conn.Exec(ctx, "INSERT INTO users (uid, name, age, email, pass) VALUES ($1, $2, $3, $4, $5)",
-		user.UID, user.Name, user.Age, user.Email, user.Pass,
-	)
-	if err != nil {
-		return err
-	}
-
+	log.Println("Database migrated successfully")
 	return nil
 }
